+++ lib/krane/cli/deploy_command.rb
@@ -25,9 +25,9 @@ module Krane
                      default: true },
         "selector" => { type: :string, banner: "'label=value'",
                         desc: "Select workloads by selector(s)" },
-        "select-any" => { type: :boolean,
-                          desc: "Enable selecting a subset of resources to deploy without validation failure",
-                          default: false },
+        "selector-as-filter" => { type: :boolean,
+                                  desc: "Use --selector as a label filter to select a subset of resources to deploy",
+                                  default: false },
         "verbose-log-prefix" => { type: :boolean, desc: "Add [context][namespace] to the log prefix",
                                   default: false },
         "verify-result" => { type: :boolean, default: true,
@@ -40,7 +40,7 @@ module Krane
         require 'krane/label_selector'
 
         selector = ::Krane::LabelSelector.parse(options[:selector]) if options[:selector]
-        select_any = options['select-any']
+        selector_as_filter = options['selector-as-filter']
 
         logger = ::Krane::FormattedLogger.build(namespace, context,
           verbose_prefix: options['verbose-log-prefix'])
@@ -64,7 +64,7 @@ module Krane
             logger: logger,
             global_timeout: ::Krane::DurationParser.new(options["global-timeout"]).parse!.to_i,
             selector: selector,
-            select_any: select_any,
+            selector_as_filter: selector_as_filter,
             protected_namespaces: protected_namespaces,
           )