+++ test/exe/deploy_test.rb
@@ -140,7 +140,7 @@ class DeployTest < Krane::TestCase
         logger: logger,
         global_timeout: 300,
         selector: nil,
-        select_any: false,
+        selector_as_filter: false,
         protected_namespaces: ["default", "kube-system", "kube-public"],
       }.merge(new_args),
       run_args: {